requirements.txt
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
PyYAML>=6.0.0
albumentations>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.0 